added PartitionedWriter
