+++ polars/polars-io/src/partition.rs
@@ -1,0 +1,170 @@
+use crate::{utils::resolve_homedir, WriterFactory};
+use polars_core::prelude::*;
+use rayon::iter::IndexedParallelIterator;
+use std::{
+    fs::File,
+    io::BufWriter,
+    marker::PhantomData,
+    path::{Path, PathBuf},
+};
+
+/// partition_df must be created by the same way of partition_by
+fn resolve_partition_dir<I, S>(rootdir: &Path, by: I, partition_df: &DataFrame) -> PathBuf
+where
+    I: IntoIterator<Item = S>,
+    S: AsRef<str>,
+{
+    let mut path = PathBuf::new();
+    path.push(resolve_homedir(rootdir));
+
+    for key in by.into_iter() {
+        let value = partition_df[key.as_ref()].get(0).to_string();
+        path.push(format!("{}={}", key.as_ref(), value))
+    }
+    path
+}
+
+pub struct PartitionedWriter<F, P, I, S> {
+    option: F,
+    rootdir: P,
+    by: I,
+    parallel: bool,
+    phantom: PhantomData<S>,
+}
+
+impl<F, P, I, S> PartitionedWriter<F, P, I, S>
+where
+    F: WriterFactory + Send + Sync,
+    P: Into<PathBuf>,
+    I: IntoIterator<Item = S>,
+    S: AsRef<str>,
+{
+    pub fn new(option: F, rootdir: P, by: I) -> Self {
+        Self {
+            option,
+            rootdir,
+            by,
+            parallel: true,
+            phantom: PhantomData,
+        }
+    }
+
+    pub fn finish(self, df: &DataFrame) -> Result<()> {
+        use polars_core::POOL;
+        use rayon::iter::IntoParallelIterator;
+        use rayon::iter::ParallelIterator;
+
+        let rootdir: PathBuf = self.rootdir.into();
+        let by: Vec<String> = self.by.into_iter().map(|x| x.as_ref().to_owned()).collect();
+        let group = df.groupby(&by)?;
+
+        if self.parallel {
+            POOL.install(|| {
+                group
+                    .get_groups()
+                    .idx_ref()
+                    .into_par_iter()
+                    .enumerate()
+                    .map(|(i, t)| {
+                        let mut partition_df =
+                            unsafe { df.take_iter_unchecked(t.1.iter().map(|i| *i as usize)) };
+                        let mut path = resolve_partition_dir(&rootdir, &by, &partition_df);
+                        std::fs::create_dir_all(&path)?;
+
+                        path.push(format!(
+                            "data-{:04}.{}",
+                            i,
+                            self.option.extension().display()
+                        ));
+
+                        let file = std::fs::File::create(path)?;
+                        let writer = BufWriter::new(file);
+
+                        self.option
+                            .create_writer::<BufWriter<File>>(writer)
+                            .finish(&mut partition_df)
+                    })
+                    .collect::<Result<Vec<_>>>()
+            })?;
+        } else {
+            for (i, t) in group.get_groups().idx_ref().iter().enumerate() {
+                let mut partition_df =
+                    unsafe { df.take_iter_unchecked(t.1.iter().map(|i| *i as usize)) };
+                let mut path = resolve_partition_dir(&rootdir, &by, &partition_df);
+                std::fs::create_dir_all(&path)?;
+
+                path.push(format!(
+                    "data-{:04}.{}",
+                    i,
+                    self.option.extension().display()
+                ));
+
+                let file = std::fs::File::create(path)?;
+                let writer = BufWriter::new(file);
+
+                self.option
+                    .create_writer::<BufWriter<File>>(writer)
+                    .finish(&mut partition_df)?;
+            }
+        }
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    #[cfg(feature = "ipc")]
+    fn test_ipc_partition() -> Result<()> {
+        use crate::ipc::IpcReader;
+        use crate::SerReader;
+        use std::{io::BufReader, path::PathBuf};
+
+        use tempdir::TempDir;
+
+        use crate::prelude::IpcWriterOption;
+
+        let tempdir = TempDir::new("ipc-partition")?;
+
+        let df = df!("a" => [1, 1, 2, 3], "b" => [2, 2, 3, 4], "c" => [2, 3, 4, 5]).unwrap();
+        let by = ["a", "b"];
+        let rootdir = tempdir.path();
+
+        let option = IpcWriterOption::new();
+
+        PartitionedWriter::new(option, &rootdir, by).finish(&df)?;
+
+        let expected_dfs = [
+            df!("a" => [1, 1], "b" => [2, 2], "c" => [2, 3])?,
+            df!("a" => [2], "b" => [3], "c" => [4])?,
+            df!("a" => [3], "b" => [4], "c" => [5])?,
+        ];
+
+        let expected: Vec<(PathBuf, DataFrame)> = ["a=1/b=2", "a=2/b=3", "a=3/b=4"]
+            .into_iter()
+            .zip(expected_dfs.into_iter())
+            .map(|(p, df)| (PathBuf::from(rootdir.join(p)), df))
+            .collect();
+
+        for (expected_dir, expected_df) in expected.iter() {
+            assert!(expected_dir.exists());
+
+            let ipc_paths = std::fs::read_dir(&expected_dir)?
+                .map(|e| {
+                    let entry = e?;
+                    Ok(entry.path())
+                })
+                .collect::<Result<Vec<_>>>()?;
+
+            assert_eq!(ipc_paths.len(), 1);
+            let reader = BufReader::new(std::fs::File::open(&ipc_paths[0])?);
+            let df = IpcReader::new(reader).finish()?;
+            assert!(expected_df.frame_equal(&df));
+        }
+
+        Ok(())
+    }
+}
